pyproject.toml
[project]
name = "DYSCO"
description = ""
version = "0.0.1"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "accelerate>=1.4.0",
    "bitsandbytes>=0.45.3",
    "datasets>=3.4.1",
    "einops>=0.8.1",
    "fiftyone>=1.3.2",
    "gitpython>=3.1.44",
    "hydra-colorlog>=1.2.0",
    "hydra-core>=1.3.2",
    "hydra-optuna-sweeper>=1.2.0",
    "lightning>=2.5.0.post0",
    "numpy==1.26.0",
    "open-clip-torch>=2.31.0",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "prettytable>=3.15.1",
    "pycocotools>=2.0.8",
    "pyrootutils>=1.0.4",
    "rich>=13.9.4",
    "scienceplots>=2.1.1",
    "seaborn>=0.13.2",
    "sentence-transformers>=3.4.1",
    "sentencepiece>=0.2.0",
    "torch>=2.6.0",
    "torchmetrics>=1.6.2",
    "torchvision>=0.21.0",
    "transformers",
    "wandb>=0.19.8",
]

[project.optional-dependencies]
dev = ["pre-commit", "ipython", "black", "isort", "flake8"]

[tool.black]
line-length = 99

[tool.isort]
profile = "black"
filter_files = true

[tool.flake8]
ignore = ["E203", "E402", "E501", "F401", "F841", "RST2", "RST301"]
exclude = ["data/*","logs/*"]

[tool.pyright]
exclude = ["data/*","logs/*",".venv/*","libs/*"]

[tool.uv.sources]
transformers = { git = "https://github.com/huggingface/transformers", rev = "v4.49.0-SigLIP-2" }
